requirements.txt
python-telegram-bot
rule34
basc-py4chan
